+++ src/components/shared/navigationBars/index.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Header from './header';
+import NavigationBars from './index';
 import SignInHeader from './signInHeader';
 import TopBar from './topBar';
 import routes from '../../../constants/routes';
@@ -14,19 +14,19 @@ describe('Header', () => {
 
   it('should render SignInHeader if prop.isSigninFlow', () => {
     props.isSigninFlow = true;
-    const wrapper = shallow(<Header {...props} />);
+    const wrapper = shallow(<NavigationBars {...props} />);
     expect(wrapper).toContainMatchingElement(SignInHeader);
   });
 
   it('should render TopBar if prop.isSigninFlow', () => {
     props.isSigninFlow = false;
-    const wrapper = shallow(<Header {...props} />);
+    const wrapper = shallow(<NavigationBars {...props} />);
     expect(wrapper).toContainMatchingElement(TopBar);
   });
 
   it('should empty render if on termsOfUse page', () => {
     props.location.pathname = routes.termsOfUse.path;
-    const wrapper = shallow(<Header {...props} />);
+    const wrapper = shallow(<NavigationBars {...props} />);
     expect(wrapper).toBeEmptyRender();
   });
 });
