+++ src/components/shared/navigationBars/topBar/autologout/renderer.js
